refactor: (#47) ErrorResponseBuilder 사용으로 예외 처리 로직 중복 제거
